saude_inteligente_backend_starter/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.7.1
python-dotenv==1.0.1
